Implemented Import Endpoint, did some refactoring
Refactored MapNode so that there is a separate function for turning csv strings into map nodes. The regular readCsv function works exactly the same. Also, added a function to strip map nodes of its neighbors, so it can be inserted into the database (MapNode.dropNeighbors). For the import endpoint, the first file has to be the nodes csv and the second one has to be the edges csv. They have to be send using multipart form data with the name 'csvFiles'